requirements.txt
django
djangorestframework

black
pytest
pylint
flake8
coverage